backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
apscheduler==3.10.4
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0 